requirements-deepseek.txt
# Modus 3: DeepSeek OCR 2
python-pptx>=0.6.23
Pillow>=10.0.0
pdf2image>=1.17.0

# PyTorch + Transformers
torch>=2.6.0
transformers==4.46.3
tokenizers==0.20.3
einops
addict
easydict

# Flash Attention (optional, aber empfohlen)
# pip install flash-attn==2.7.3 --no-build-isolation

# 4-bit Quantisierung (optional)
# bitsandbytes

# Alternative: vLLM statt Transformers (schneller)
# pip install -U vllm --pre --extra-index-url https://wheels.vllm.ai/nightly
